various
- iid changes from integer to long
- host name changed to mdns derived
- add results to thing actions
- trigger child discovery faster
- refactor chunked decoding; add test
- tweak start up sequencing

+++ bundles/org.openhab.binding.homekit/src/main/java/org/openhab/binding/homekit/internal/dto/Accessory.java
@@ -186,7 +186,7 @@ public class Accessory {
         return toString();
     }
 
-    public @Nullable Service getService(Integer iid) {
+    public @Nullable Service getService(Long iid) {
         return services.stream().filter(s -> iid.equals(s.iid)).findFirst().orElse(null);
     }
